Cruft cleanup stage 3....
Found some refs to 'unsigned long'- which changes intent and meannig on
64-bits.  In 32-bits (which is hinted at by using 32-bit string
formatter fields...) long and int are interchangeable and were
improperly munged up by devs all over the place.  Clear int/uint32_t
type values.  Moved to int as suggested by cppcheck tests.

+++ src/lib/unix/x86cpuid.c
@@ -9,8 +9,8 @@ void nope()
 
 int main()
 {
-  unsigned long x[4];
-  unsigned long y[4];
+  unsigned int x[4];
+  unsigned int y[4];
   int i;
   int j;
   char c;
